[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17716 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -2,11 +2,11 @@ import java.util.Scanner;
 interface Main {
     public static void main(String args[]) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         
-        int plus = a + b;
-        int minus = a - b;
+        long plus = a + b;
+        long minus = a - b;
         
         System.out.println(plus * minus);
     }
